Split b tree structure into modules
Use get_settings method instead of explicit call of degrees
